Moving all activations to novi.client.activations
Adding activations for OR and AND combinations

Cleaning imports

Do not modify context as it affects activations downstream.

Adding a register decorator to provide a 2nd way of registering activations.

Renaming all models to have the suffix model to distinguish from the activation business objects.

Adding methods to retrieve activations
Rename method

Adding new endpoint evaluatedFlags instead of query param

Adding logging for the web app.
Ability to pass a script to seed database.

Adding logging for the web app.